pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "ngff-zarr"
description = 'A lean and kind Open Microscopy Environment (OME) Next Generation File Format (NGFF) Zarr implementation.'
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Matt McCormick", email = "matt.mccormick@kitware.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy",
    "zarr",
    "dask[array]",
    "platformdirs",
    "rich",
    "typing_extensions",
    "psutil; sys_platform != \"emscripten\""
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/thewtex/ngff-zarr#readme"
Issues = "https://github.com/thewtex/ngff-zarr/issues"
Source = "https://github.com/thewtex/ngff-zarr"
Changelog = "https://github.com/thewtex/ngff-zarr/releases"

[project.scripts]
ngff-zarr = "ngff_zarr.cli:main"

[tool.hatch.version]
path = "ngff_zarr/__about__.py"

[project.optional-dependencies]
dask-image = [
    "dask-image",
]
itk = [
    "itk-filtering>=5.3.0",
]
cli = [
    "dask-image",
    "dask[distributed]",
    "itk-filtering>=5.3.0",
    "itk-io>=5.3.0",
    "imageio",
    "tifffile",
    "imagecodecs",
]
test = [
    "pytest >=6",
    "pre-commit",
    "pooch",
    "itkwasm",
    "itk-io>=5.3.0",
    "itk-filtering>=5.3.0",
    "tifffile",
]


[tool.coverage.report]
run.source = ["ngff_zarr"]
report.exclude_lines = [
  "no cov",
  '\.\.\.',
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]


[tool.pytest.ini_options]
minversion = "6.0"
filterwarnings = [
  "ignore:(ast.Str|Attribute s|ast.NameConstant|ast.Num) is deprecated:DeprecationWarning:_pytest",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]


[tool.ruff]
select = [
  "E", "F", "W", # flake8
  "B",           # flake8-bugbear
  "I",           # isort
  "ARG",         # flake8-unused-arguments
  "C4",          # flake8-comprehensions
  "EM",          # flake8-errmsg
  "ICN",         # flake8-import-conventions
  "ISC",         # flake8-implicit-str-concat
  "G",           # flake8-logging-format
  "PGH",         # pygrep-hooks
  "PIE",         # flake8-pie
  "PL",          # pylint
  "PT",          # flake8-pytest-style
  "PTH",         # flake8-use-pathlib
  "RET",         # flake8-return
  "RUF",         # Ruff-specific
  "SIM",         # flake8-simplify
  "T20",         # flake8-print
  "UP",          # pyupgrade
  "YTT",         # flake8-2020
  "EXE",         # flake8-executable
  "NPY",         # NumPy specific rules
  "PD",          # pandas-vet
]
extend-ignore = [
  "PLR",    # Design related pylint codes
  "E501",   # Line too long
]
src = ["src"]
unfixable = [
  "T20",  # Removes print statements
  "F841", # Removes unused variables
]
exclude = []
flake8-unused-arguments.ignore-variadic-names = true
line-length = 88

[tool.ruff.per-file-ignores]
"tests/**" = ["T20"]


[tool.pylint]
py-version = "3.8"
ignore-paths= ["src/ngff_zarr/__about__.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "wrong-import-position",
]

[tool.black]
line-length = 88
